Added LifeCycle
