+++ src/electron/output/helpers/OutputHelper.ts
@@ -4,12 +4,15 @@ import { OutputIdentify } from "./OutputIdentify"
 import { OutputSend } from "./OutputSend"
 import { OutputValues } from "./OutputValues"
 import { OutputVisibility } from "./OutputVisibility"
+import { OutputLifecycle } from "./OutputLifecycle"
 
 export class OutputHelper {
     static outputWindows: { [key: string]: BrowserWindow } = {}
-    static Identify = OutputIdentify
-    static Send = OutputSend
+
     static Bounds = OutputBounds
+    static Identify = OutputIdentify
+    static Lifecycle = OutputLifecycle
+    static Send = OutputSend
     static Values = OutputValues
     static Visibility = OutputVisibility
 }
